+++ src/components/views/Sidebar.jsx
@@ -9,13 +9,14 @@ import Badge from 'apollo-react/components/Badge';
 import PlusIcon from 'apollo-react-icons/Plus';
 import CardIcon from 'apollo-react-icons/Card';
 import SyncIcon from 'apollo-react-icons/Sync';
+import Download from 'apollo-react-icons/Download';
 import Close from 'apollo-react-icons/Close';
 import IconButton from 'apollo-react/components/IconButton';
 import Typography from 'apollo-react/components/Typography';
 import Tooltip from 'apollo-react/components/Tooltip';
 import { neptunePrimaryDark, neutral7 } from 'apollo-react/colors';
 
-import Notepad from './Notepad';
+// import Notepad from './Notepad';
 import chevronRight from '../../../img/chevron-right.svg';
 import {
   handleSelectedSection,
@@ -33,6 +34,7 @@ import { REDUX_TYPES } from '../../constants';
 
 import MatomoHOC from '../HOC/MatomoHOC';
 import { selectAreAllSectionsExpanded } from '../../redux/selectors/proposal';
+import {actionChannel, UI_ACTION} from '../../uiActions/ui-actions'
 const MANUAL_REFRESH = false;
 
 type Props = {
@@ -244,18 +246,6 @@ class Sidebar extends Component<Props, State> {
       selectedBid
     } = this.props;
     const { selectedSection, activeTabIndex } = this.state;
-    const NotepadTab = () =>
-      notes.size === 0 ? (
-        <Typography variant="body2" style={{ fontWeight: 'inherit' }}>
-          Notepad
-        </Typography>
-      ) : (
-        <Badge variant="dot">
-          <Typography variant="body2" style={{ fontWeight: 'inherit' }}>
-            Notepad
-          </Typography>
-        </Badge>
-      );
 
     return (
       <div
@@ -320,6 +310,21 @@ class Sidebar extends Component<Props, State> {
                   }}
                 />
               </Tooltip>
+              <Tooltip title="Export Opportunity" placement="top">
+                <Download
+                  style={{
+                    color: neptunePrimaryDark,
+                    width: 20,
+                    height: 20,
+                    margin: 3,
+                    cursor: 'pointer'
+                  }}
+                  onClick={e => {
+                    this.handleItemsVisibility(e);
+                    actionChannel.next({ name : UI_ACTION.openGenerateModal })
+                  }}
+                />
+              </Tooltip>
               { MANUAL_REFRESH &&  
               <Tooltip title="Refresh Proposal Sources" placement="right">
                 <SyncIcon
@@ -349,7 +354,6 @@ class Sidebar extends Component<Props, State> {
               truncate
             >
               <Tab label="Index" />
-              <Tab label={<NotepadTab />} style={{ paddingRight: '8px' }} />
             </Tabs>
             {activeTabIndex === 0 && (
               <div className="sidebar-content-list">
@@ -381,14 +385,6 @@ class Sidebar extends Component<Props, State> {
                   return null;
                 })}
               </div>
-            )}
-            {activeTabIndex === 1 && (
-              <Notepad
-                sections={sections}
-                id={id}
-                selectedtitle={selectedtitle || ''}
-                trackMatomoNoteSubmit={this.trackMatomoNoteSubmit}
-              />
             )}
           </div>
         </div>